requirements-minimal.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic
pydantic-settings
email-validator 